ci: bump GitHub Actions versions
actions/setup-go
The recommendation is based on the YAML parser's behavior, which interprets non-wrapped
values as numbers and, in the case of version 1.20, trims it down to 1.2, which may not be
very obvious.

actions/upload-artifact
actions/upload-artifact@v3 is scheduled for deprecation on November 30, 2024. Learn more.
Similarly, v1/v2 are scheduled for deprecation on June 30, 2024. Please update your workflow
to use v4 of the artifact actions. This deprecation will not impact any existing versions of
GitHub Enterprise Server being used by customers.