fix(main): fix staled binary fix_staled_instances
New wsl2-ssh-bridge is created for every GPG connection, but it is not
removed after the connection is closed.

After investigation in https://github.com/BlackReloaded/wsl2-ssh-pageant/issues/38
looks like issue is related to change introduced in
https://github.com/BlackReloaded/wsl2-ssh-pageant/commit/937e3d8e82ae71a73f6ec6a390e00e91776efa1a
as a fix to https://github.com/BlackReloaded/wsl2-ssh-pageant/issues/30

This commit reverts the change and fixes the issue.